made the thing work and did another thing can't type becuase I gotta e quick ahhh

+++ js/script.js
@@ -3,7 +3,7 @@
 // function to indicate wether the temperature is cold or hot
 function buttonClicked(){
   // variables for user's temperature input
-  let temp = parseInt(document.getElementById("temp").value)
+  let temp = parseFloat(document.getElementById("temp").value)
    
   // if statement that checks if the temperature is hot, cold, or just right
   if(temp < 0){
